pyproject.toml
[project]
name = "fantasy-football-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "espn-api>=0.45.1",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
]
